Remove cmake from blacklist, now that --with-docs is default for octave
I think the reason the sphinx-doc dependency was breaking before is that
the --without-docs option passed to octave in our build process was
interacting badly with the build_with?("docs") test in cmake's sphinx-doc
dependency setup. Now that we're not doing --without-docs, I think the normal
cmake formula will work again.