use "-octapp" suffix instead of "-octave-app"
The "-octapp" suffix is both more concise and more distinctive. There are other places where the string "octave-app" occurs, including the org and repo hames. And especially when saying it out loud, "octave octave app" is a real mouthful, and is ambiguous with e.g. "the octave app" when using "app" to just mean an application or Mac app bundle.