remove rolling minor-version-only octave formulae, in favor of specific patch versions
I don't think these rolling minor-version formulae are actually useful to this project, since we're always working with specific patch versions for our builds, and don't intend these formulae to really be a means of delivering octave to end users. So the rolling minor-release formulae are just another file to maintain. I waffled on this a bit with the 9.0 series, but realized it's appropriate specifically for the x.0 prerelease series, because those come out with a lot of minor versions in the leadup to the GM release.