8.x, 9.x: remove --std=gnu++14 option (sundials-oriented?), aligning with core formula
This isn't --std=gnu++14 present in core Homebrew octave formulae. I'm guessing gnu++14 is the default now with newer versions of GCC, or something like that? Or maybe it's that gnu++17 is the default now, and that's now supported by sundials and Octave, and the --std=gnu++14 was actually a downgrade to support sundials or other math libraries that didn't support gnu++17 yet?

I don't really know what I'm doing here. Just bringing things more in line with the core formulae, which I assume has been better tested. That's something I think we can do now that Octave builds against the core Homebrew sundials formula, instead of an older version.