add octave-polyalloc@9.1.0 formula, with C++ polymorphic allocator option
This is a minor variant of the vanilla (non-octapp) octave@9.1.0 formula, just with the new C++ --enable-std-pmr-polymorphic-allocator build option added. That's the only difference, but I made it a distinct formula instead of a build option on the main formula so that a) they can be installed side-by-side and compared on a single machine without re-building to swap them, and b) you can't accidentally invoke it on the main formula, and the git history reflects which options we use where.

One of the core Octave maintainers, mmuetzel I think, expressed interest in having this, on the Octave Discourse in one of these threads (though I can't find the exact comment now):

* https://octave.discourse.group/t/enable-polymorphic-allocator-by-default-if-available/5132/20
* https://octave.discourse.group/t/help-building-octave-on-macos-with-homebrew/5481/64
* https://octave.discourse.group/t/improve-access-to-octave-on-macos/4956/15
* https://octave.discourse.group/t/execution-error-on-macos-runners-with-apple-clang-15/5258/2
  * (this might be the actual request)