Remove octave-app cask, bc I'm out of date and can't support AS
The Cask setup, last I checked, doesn't support different downloads for different CPU architecture. But I also haven't really looked at the Homebrew Cask stuff in a while, and maybe just don't know how to do that. Either way, I think if I bump this to 8.4.0 it'll be a broken install in some cases, with a hard recovery path. And I don't know if we have *any* cask users. So I'm just killing the cask, at least for now.