drop custom qscintilla2-octapp in favor of vanilla qscintilla2 for Qt 6 builds
There's no longer any difference between the qscintilla2-octapp variant and the core qscintilla2 formula, so I don't see a reason to keep it around just as a a placeholder. We can recreate it later if it has to diverge again, and not have to maintain it in the mean time.