9.1.0 w qt5: remove bogus build.with?("qt") check that made it ignore Qt
In 9.1, I made qt non-optional, to bring the -octapp formulae in line with the core Homebrew formulae. Looks like I missed removing the `build.with?("qt")` checks on this formula, even though I removed the --without-qt build option, and I guess brew doesn't check to see if a tested option is actually available in the formula and treats it as not enabled. This caused the formula to build without qt enabled, and the GUI couldn't launch.