ocatpp: drop librsvg dependency, to avoid heavy rust and llvm dependencies
I don't really remember why we have a librsvg dependency. I think it's actually just a build-time dependency for icon manipulation. In which case, we can just make it a "build environment" requirement instead of including it in the octapp build itself. It has a dependency on rust (and llvm), which is heavy (long build time), so converting it to a build toolchain dependency would reduce our build times, and I'm sick of our long build times.

This also makes sure the rust and llvm packages won't end up in the final packaged app, which can inflate the app size by 2 GB or more, on top of an initial ~3 GB size.