octave@4.4.1: remove gnuplot dep, bc supporting it is too hard
I don't have a gnuplot for qt512 any more, gnuplot is hard to support and get to work with Qt in general, upstream has dropped it, and this is just the vanilla octave@4.4.1 that I seldom use. Sorry, gnuplot fans.

Even if I'm not building octave@4.4.1, the presence of the bad gnuplot-qt512 reference breaks some brew functionality, like `brew uses <formula> --eval-all --recursive`, which I currently want to use to investigate llvm@17 usage.