octapp*: bundle gnu-units package, to support miscellaneous OF package
macOS supplies a units command, but it is BSD units instead of GNU units, with different behavior and definition sets. The miscellaneous package expects GNU units; tests will fail and functionality reduced with BSD units.

The miscellaneous package's code doesn't look for the `gunits` g-prefixed variant that this package installs, but if/when it changes to do so, the package will be here waiting for it. I'm adding it preemptively now, because we're currently putting together the 9.x and 8.4u1 releaess so it's a good time to add stuff. And the package is only 1 MB and unlikely to conflict with anything (except newer versions of gnu-units in a system Homebrew, but gnu-units releases rarely, so that's probably not a problem).