add missing MacTeX requirement to octave-devel*, drop 'depends_on "texlive"'
The absence of the MacTeX requirement and add-to-path was just an oversight in my initial creation of the octave-devel* formulae. Using `depends_on "texlive"` is a viable alternative to it, but we're using MacTeX so far, so I'm sticking with that uniformly for now.

See: https://github.com/octave-app/octave-app/issues/293