Fix octave-(stable,devel) for recent changes, make octave-devel head-only
These formulae haven't been tweaked in a year or maybe two, so that's how much change this is picking up. This change got octave-stable, but not octave-devel, to build for me on macOS 15 on Intel.

Removed the --with-test to bring these in line with the mainline formulae, since I never use that anyway.