add octave@6.4.0
The vanilla Octave 6.4.0 formula got missed first time around. Adding it retroactively. It's broken - the build fails with a "sip-install: Unable to import 'project.py': No module named 'pyqtbuild'" error - but I think that's still useful information, and this formula is around as a reference if we want to fix that up some day.